[ci skip]

+++ MzLibPlugin/src/main/java/mz/mzlib/minecraft/SleepTicks.java
@@ -4,9 +4,9 @@ import mz.mzlib.util.async.BasicAwait;
 
 public class SleepTicks implements BasicAwait
 {
-    public int ticks;
+    public long ticks;
 
-    public SleepTicks(int ticks)
+    public SleepTicks(long ticks)
     {
         this.ticks = ticks;
     }
